Relocate everything to not overlap logging
