+++ yet-another-docker-its/src/test/resources/logback-test.xml
@@ -6,8 +6,8 @@
         </encoder>
     </appender>
 
-    <logger name="com.github.kostyasha.yad.docker_java.com.github.dockerjava" level="DEBUG"/>
-    <logger name="com.github.kostyasha.yad.docker_java.org.apache.http.impl.conn" level="DEBUG"/>
+    <logger name="com.github.kostyasha.yad_docker_java.com.github.dockerjava" level="DEBUG"/>
+    <logger name="com.github.kostyasha.yad_docker_java.org.apache.http.impl.conn" level="DEBUG"/>
     <logger name="com.github.kostyasha.it" level="TRACE"/>
