+++ yet-another-docker-plugin/src/main/java/com/github/kostyasha/yad/DockerCloud.java
@@ -2,17 +2,17 @@ package com.github.kostyasha.yad;
 
 import com.github.kostyasha.yad.commons.AbstractCloud;
 import com.github.kostyasha.yad.commons.DockerCreateContainer;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.DockerClient;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.command.CreateContainerCmd;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.command.CreateContainerResponse;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.command.InspectContainerResponse;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.command.StartContainerCmd;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.exception.DockerException;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.Container;
-import com.github.kostyasha.yad.docker_java.com.google.common.base.Throwables;
-import com.github.kostyasha.yad.docker_java.javax.ws.rs.ProcessingException;
-import com.github.kostyasha.yad.docker_java.org.apache.commons.lang.StringUtils;
-import com.github.kostyasha.yad.docker_java.org.apache.commons.lang.builder.ToStringBuilder;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.DockerClient;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.command.CreateContainerCmd;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.command.CreateContainerResponse;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.command.InspectContainerResponse;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.command.StartContainerCmd;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.exception.DockerException;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.model.Container;
+import com.github.kostyasha.yad_docker_java.com.google.common.base.Throwables;
+import com.github.kostyasha.yad_docker_java.javax.ws.rs.ProcessingException;
+import com.github.kostyasha.yad_docker_java.org.apache.commons.lang.StringUtils;
+import com.github.kostyasha.yad_docker_java.org.apache.commons.lang.builder.ToStringBuilder;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import hudson.Extension;
 import hudson.model.Computer;
