+++ yet-another-docker-plugin/src/main/java/com/github/kostyasha/yad/DockerConnector.java
@@ -3,13 +3,13 @@ package com.github.kostyasha.yad;
 import com.cloudbees.plugins.credentials.CredentialsMatchers;
 import com.cloudbees.plugins.credentials.CredentialsProvider;
 import com.cloudbees.plugins.credentials.common.StandardCredentials;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.DockerClient;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.exception.DockerException;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.Version;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.DefaultDockerClientConfig;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.RemoteApiVersion;
-import com.github.kostyasha.yad.docker_java.com.google.common.base.Preconditions;
-import com.github.kostyasha.yad.docker_java.org.apache.commons.lang.StringUtils;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.DockerClient;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.exception.DockerException;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.model.Version;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.DefaultDockerClientConfig;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.RemoteApiVersion;
+import com.github.kostyasha.yad_docker_java.com.google.common.base.Preconditions;
+import com.github.kostyasha.yad_docker_java.org.apache.commons.lang.StringUtils;
 import com.github.kostyasha.yad.other.ConnectorType;
 import com.github.kostyasha.yad.utils.CredentialsListBoxModel;
 import com.google.common.base.Throwables;
@@ -37,7 +37,7 @@ import java.util.Collections;
 import java.util.List;
 
 import static com.github.kostyasha.yad.client.ClientBuilderForConnector.newClientBuilderForConnector;
-import static com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.RemoteApiVersion.parseConfig;
+import static com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.RemoteApiVersion.parseConfig;
 import static com.github.kostyasha.yad.other.ConnectorType.NETTY;
 import static hudson.util.FormValidation.ok;
 import static hudson.util.FormValidation.warning;
