+++ yet-another-docker-plugin/src/main/java/com/github/kostyasha/yad/commons/DockerPullImage.java
@@ -6,13 +6,13 @@ import com.cloudbees.plugins.credentials.CredentialsProvider;
 import com.cloudbees.plugins.credentials.common.StandardListBoxModel;
 import com.cloudbees.plugins.credentials.domains.DomainRequirement;
 import com.github.kostyasha.yad.credentials.DockerRegistryAuthCredentials;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.DockerClient;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.command.PullImageCmd;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.Image;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.NameParser;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.command.PullImageResultCallback;
-import com.github.kostyasha.yad.docker_java.com.google.common.collect.Iterables;
-import com.github.kostyasha.yad.docker_java.org.apache.commons.lang.StringUtils;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.DockerClient;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.command.PullImageCmd;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.model.Image;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.NameParser;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.command.PullImageResultCallback;
+import com.github.kostyasha.yad_docker_java.com.google.common.collect.Iterables;
+import com.github.kostyasha.yad_docker_java.org.apache.commons.lang.StringUtils;
 import hudson.Extension;
 import hudson.model.AbstractDescribableImpl;
 import hudson.model.Descriptor;
