+++ yet-another-docker-plugin/src/main/java/com/github/kostyasha/yad/commons/DockerStopContainer.java
@@ -1,7 +1,7 @@
 package com.github.kostyasha.yad.commons;
 
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.DockerClient;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.command.StopContainerCmdImpl;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.DockerClient;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.command.StopContainerCmdImpl;
 import hudson.Extension;
 import hudson.model.AbstractDescribableImpl;
 import hudson.model.Descriptor;
