+++ yet-another-docker-plugin/src/main/java/com/github/kostyasha/yad/launcher/DockerComputerSSHLauncher.java
@@ -3,15 +3,15 @@ package com.github.kostyasha.yad.launcher;
 import com.github.kostyasha.yad.DockerCloud;
 import com.github.kostyasha.yad.DockerSlaveTemplate;
 import com.github.kostyasha.yad.commons.DockerCreateContainer;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.command.CreateContainerCmd;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.command.InspectContainerResponse;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.ExposedPort;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.NetworkSettings;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.PortBinding;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.model.Ports;
-import com.github.kostyasha.yad.docker_java.com.google.common.annotations.Beta;
-import com.github.kostyasha.yad.docker_java.com.google.common.base.Preconditions;
-import com.github.kostyasha.yad.docker_java.com.google.common.net.HostAndPort;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.command.CreateContainerCmd;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.command.InspectContainerResponse;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.model.ExposedPort;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.model.NetworkSettings;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.model.PortBinding;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.model.Ports;
+import com.github.kostyasha.yad_docker_java.com.google.common.annotations.Beta;
+import com.github.kostyasha.yad_docker_java.com.google.common.base.Preconditions;
+import com.github.kostyasha.yad_docker_java.com.google.common.net.HostAndPort;
 import com.github.kostyasha.yad.utils.HostAndPortChecker;
 import hudson.model.Descriptor;
 import hudson.model.ItemGroup;
