+++ yet-another-docker-plugin/src/main/java/com/github/kostyasha/yad/other/VariableSSLConfig.java
@@ -1,9 +1,9 @@
 package com.github.kostyasha.yad.other;
 
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.api.exception.DockerClientException;
-import com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.SSLConfig;
-import com.github.kostyasha.yad.docker_java.org.bouncycastle.jce.provider.BouncyCastleProvider;
-import com.github.kostyasha.yad.docker_java.org.glassfish.jersey.SslConfigurator;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.api.exception.DockerClientException;
+import com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.SSLConfig;
+import com.github.kostyasha.yad_docker_java.org.bouncycastle.jce.provider.BouncyCastleProvider;
+import com.github.kostyasha.yad_docker_java.org.glassfish.jersey.SslConfigurator;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 import javax.net.ssl.SSLContext;
@@ -14,8 +14,8 @@ import java.security.NoSuchAlgorithmException;
 import java.security.Security;
 import java.security.UnrecoverableKeyException;
 
-import static com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.util.CertificateUtils.createKeyStore;
-import static com.github.kostyasha.yad.docker_java.com.github.dockerjava.core.util.CertificateUtils.createTrustStore;
+import static com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.util.CertificateUtils.createKeyStore;
+import static com.github.kostyasha.yad_docker_java.com.github.dockerjava.core.util.CertificateUtils.createTrustStore;
 import static java.util.Objects.nonNull;
 import static org.apache.commons.lang.StringUtils.isNotBlank;
